# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ELOJIM API
  description: API oficial del proyecto Fundación ELOJIM
  version: "1.0.0"
  contact:
    name: Soporte ELOJIM
    email: soporte@elojim.org
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    RegistroBase:
      type: object
      properties:
        nombreCompleto:
          type: string
          example: "Laura Gómez"
        fechaNacimiento:
          type: string
          format: date
        edad:
          type: integer
        comuna:
          type: string
        estratoSocial:
          type: string
        grupoEtnico:
          type: string
        telefonoContacto:
          type: string
        correoElectronico:
          type: string
        direccion:
          type: string
        motivacion:
          type: string
        expectativas:
          type: string
        aceptaTerminos:
          type: boolean
      required:
        - nombreCompleto
        - fechaNacimiento
        - comuna
        - estratoSocial
        - telefonoContacto
        - correoElectronico
        - aceptaTerminos
...
    RegistroCultural:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            documentoIdentidad:
              type: string
            municipioDepartamento:
              type: string
            nivelEducativo:
              type: string
            ocupacion:
              type: string
            areaInteresPrincipal:
              type: string
            formacionPrevia:
              type: boolean
            detalleFormacion:
              type: string
            perteneceGrupo:
              type: boolean
            detalleGrupo:
              type: string
            diasDisponibles:
              type: string
          example:
            nombreCompleto: "Dylan Ramirez"
            areaInteresPrincipal: "Musica"
            formacionPrevia: true
            perteneceGrupo: true
            aceptaTerminos: true
...
security:
  - cookieAuth: []
paths:
...
  /api/inscripciones-evento:
    post:
      tags: [Eventos]
      summary: Inscribirse a un evento
      description: Requiere estar autenticado con sesión de NextAuth
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [programId, eventoId]
              properties:
                programId:
                  type: string
                  example: mujer-vulnerable
                eventoId:
                  type: integer
                  example: 12
      responses:
        "201": { description: Inscripción realizada con éxito }
        "401": { description: No autenticado }
  /api/inscripciones-evento/{eventId}:
    get:
      tags: [Eventos]
      summary: Listar inscripciones a un evento
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lista de inscripciones
  
  # -------------------------
  # NEWS
  # -------------------------
...
  /api/registro/cultural:
    get:
      tags: [Registro]
      summary: Listar registros del programa cultural
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro para Cultural
      description: Crear un nuevo registro para el programa Cultural.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroCultural"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de programa cultural
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno      
...